HiddenMarkovModelReaders

Documentation for HiddenMarkovModelReaders.

HiddenMarkovModelReaders.HMM

HiddenMarkovModelReaders.HMMParams

HiddenMarkovModelReaders.ScorePair

HiddenMarkovModelReaders.amplitude

HiddenMarkovModelReaders.backTrace

HiddenMarkovModelReaders.bhattDist

HiddenMarkovModelReaders.euclDist

HiddenMarkovModelReaders.feed!

HiddenMarkovModelReaders.process!

HiddenMarkovModelReaders.reset!

HiddenMarkovModelReaders.setup

HiddenMarkovModelReaders.sortHMM!

HiddenMarkovModelReaders.HMM — Type

Hidden Markov model object.

Fields

data::Array: hold data
model::Array: keep model scores
traceback::Array: collect traceback

See also: HMMParams, setup,

HiddenMarkovModelReaders.HMMParams — Type

Hidden Markov model parameters.

Fields

penalty::Float64: penalty on hidden Markov model. Defualt = 200.
minimumFrequency::Int64: minimun frequency on model state to split during hypothesis generation. Default = 20.
verbosity::Bool: increase verbosity. Default = false.
distance::Function: calculate distance. Possible options provided by this package inlcude: `euclDist`, `bhattDist`. Alternatively, a predefined or an anonymous / λ functions with the form `functionDist(arr::Array{T, 1}, h::Array{T, 1}) where T <: Number` can be passed.

See also: HMM, setup, euclDist, bhattDist,

HiddenMarkovModelReaders.ScorePair — Type

Keep track of scores & indexes.

Fields

score::Float64: score
index::Int64: index

See also: HMM, HMMParams

HiddenMarkovModelReaders.amplitude — Method

amplitude(arr::Array{T, 1}) where {T <: Number}

Description

Amplitude.

Examples

julia> amplitude(collect(1:10))
19.621416870348583

See also: euclDist, bhattDist

HiddenMarkovModelReaders.backTrace — Method

Backtrace on hidden Markov model object.

HiddenMarkovModelReaders.bhattDist — Method

bhattDist(arr::Array{T, 1}, h::Array{T, 1}) where {T <: Number}

Description

Bhattacharyya distance.

Examples

julia> bhattDist(collect(1:10), fill(5, 10))
-3.936532135073928

See also: euclDist, amplitude

HiddenMarkovModelReaders.euclDist — Method

euclDist(arr::Array{T, 1}, h::Array{T, 1}) where {T <: Number}

Description

Euclidean distance.

Examples

julia> euclDist(collect(1:10), fill(5, 10))
9.219544457292887

See also: bhattDist, amplitude

HiddenMarkovModelReaders.feed! — Method

Feed on hidden Markov model object.

HiddenMarkovModelReaders.process! — Method

process(self::HMM, d::Array{T, 2}, splitSw::Bool;
params::HMMParams) where T <: Number

Description

Process hidden Markov model object. Meant as an iterative mutating function, perform several steps:

reset model traceback.

feed frames into model data.

update model.

generate hypothesized new states.

See also: setup, HMM, HMMParams

HiddenMarkovModelReaders.reset! — Method

Reset hidden Markov model object.

HiddenMarkovModelReaders.setup — Method

setup(v::Array{T, 2}) where <: Number

Initiate hidden Markov model object.

See also: process!, HMM, HMMParams

HiddenMarkovModelReaders.sortHMM! — Method

sort hidden Markov model by amplitude.